Coxeter Groups

This module constructs finite Coxeter groups as permutation groups acting on their root systems, and provides algorithms for lengths, reduced words, reflections, parabolic subgroups, and conjugacy classes.

Building a Coxeter Group

OrbitAl.coxeter.coxeterGraph — Function

coxeterGraph(series, rank)

Return the edge list of the Coxeter diagram for the group of given series ("A", "B", "C", "D", "E", "F", ...) and rank. Each edge (i, j) indicates that generators s_i and s_j have Coxeter order 3.

Examples

julia> using OrbitAl

julia> coxeterGraph("A", 3)
2-element Vector{Tuple{Int64, Int64}}:
 (1, 2)
 (2, 3)

OrbitAl.coxeter.coxeterMat — Function

coxeterMat(series, rank)

Return the Coxeter matrix for the group of given series and rank. Entry m[i,j] is the order of the product s_i * s_j of generators.

Examples

julia> using OrbitAl

julia> coxeterMat("A", 2)
2×2 Matrix{Int64}:
 1  3
 3  1

OrbitAl.coxeter.cartanMat — Function

cartanMat(series, rank)

Return the Cartan matrix for the Lie algebra of given series and rank.

Examples

julia> using OrbitAl

julia> cartanMat("A", 2)
2×2 Matrix{Int64}:
  2  -1
 -1   2

OrbitAl.coxeter.CoxeterGp — Type

CoxeterGp(C)

Construct a Coxeter group from the Cartan matrix C, realized as a permutation group acting on its root system. The generators correspond to simple reflections.

Examples

julia> using OrbitAl

julia> W = CoxeterGp(cartanMat("A", 2));

julia> sizeOfGroup(W)
6

Length and Words

OrbitAl.coxeter.coxeterLength — Function

coxeterLength(W, w)

Return the Coxeter length of w in W, i.e., the number of positive roots sent to negative roots by w, which equals the length of a shortest generator word for w.

Examples

julia> using OrbitAl

julia> W = CoxeterGp(cartanMat("A", 2));

julia> coxeterLength(W, W.gens[1] * W.gens[2])
2

OrbitAl.coxeter.coxeterWord — Function

coxeterWord(W, w)

Return a reduced word for w in W as a vector of generator indices, computed by iteratively stripping left descents.

Examples

julia> using OrbitAl

julia> W = CoxeterGp(cartanMat("A", 2));

julia> length(coxeterWord(W, W.gens[1] * W.gens[2]))
2

OrbitAl.coxeter.permCoxeterWord — Function

permCoxeterWord(W, word)

Return the permutation in W corresponding to the word word, a vector of generator indices.

OrbitAl.coxeter.reflections — Function

reflections(W)

Return all reflections in the Coxeter group W as a vector of permutations. Reflections are conjugates of the generators; their count equals the number of positive roots.

Examples

julia> using OrbitAl

julia> W = CoxeterGp(cartanMat("A", 2));

julia> length(reflections(W))
3

Parabolic Subgroups

OrbitAl.coxeter.longestElt — Function

longestElt(W, J)

Return the longest element of the parabolic subgroup of W generated by the simple reflections indexed by J.

OrbitAl.coxeter.parabolicTransversal — Function

parabolicTransversal(W, J)

Return the shortest left coset representatives for the parabolic subgroup of W generated by J.

OrbitAl.coxeter.prefixes — Function

prefixes(W, w)

Return all right-sided prefixes of w in the right weak order of W, i.e., the elements obtained from w by removing right descents one at a time.

OrbitAl.coxeter.prefixes_with_edges — Function

prefixes_with_edges(W, w)

Like prefixes, but also returns the labeled edges of the Hasse diagram of the right weak order interval below w.

Shapes and Conjugacy

OrbitAl.coxeter.shape — Function

shape(W, J)

Return the orbit of the parabolic subset J under conjugation by double coset representatives, i.e., the set of all parabolic subsets conjugate to J in W.

OrbitAl.coxeter.shapes — Function

shapes(W)

Return all parabolic shapes in W, i.e., the orbits of all parabolic subsets under conjugation by double coset representatives.

OrbitAl.coxeter.coxeterConjugacyClasses — Function

coxeterConjugacyClasses(W)

Return a list of minimum-length conjugacy class representatives in W, one per conjugacy class.