Permutation Groups

This module provides lightweight permutation groups built on top of the orbit algorithms. A group is defined by a list of generators and an identity element; all group-theoretic operations (size, membership, conjugacy classes) are derived from orbit computations.

Types

OrbitAl.permgroup.PermGp — Type

PermGp(gens, one)

A permutation group with generators gens::Vector{Perm} and identity element one.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> G = PermGp([s, t], one(s));

julia> sizeOfGroup(G)
6

Element Enumeration

OrbitAl.permgroup.elements — Function

elements(group)

Return all elements of group as a sorted vector.

OrbitAl.permgroup.subgroups — Function

subgroups(group)

Return all subgroups of group as a vector of PermGp values, computed by orbit enumeration under the closure action.

Size and Membership

OrbitAl.permgroup.sizeOfGroup — Function

sizeOfGroup(group)

Return the order of group using the orbit-stabilizer theorem, without enumerating all elements.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> sizeOfGroup(PermGp([s, t], one(s)))
6

OrbitAl.permgroup.memberOfGroup — Function

memberOfGroup(group, perm)

Return true if perm belongs to group, using a recursive sifting algorithm.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> G = PermGp([s, t], one(s));

julia> memberOfGroup(G, s)
true

julia> memberOfGroup(G, Perm([2,3,4,1]))
false

OrbitAl.permgroup.randomGroupElement — Function

randomGroupElement(group)

Return a uniformly random element of group using a recursive orbit-stabilizer decomposition.

Conjugacy

OrbitAl.permgroup.conjClasses — Function

conjClasses(gp)

Return the conjugacy classes of gp as a vector of Orbit objects, each containing the elements of one class.

OrbitAl.permgroup.subgpClasses — Function

subgpClasses(gp)

Return the conjugacy classes of subgroups of gp as a vector of Orbit objects.

OrbitAl.permgroup.closure — Function

closure(group, a)

Return a new PermGp formed by adding generator a to group.

Subgroup Structure

OrbitAl.permgroup.isPrimePower — Function

isPrimePower(n)

Return true if n > 1 is a power of a single prime, i.e., n = p^k for some prime p and k ≥ 1.

OrbitAl.permgroup.zuppos — Function

zuppos(group)

Return one generator for each cyclic subgroup of prime-power order (zuppo) of group. The returned generators suffice to find all subgroups of group under the closure action.

Intersection

OrbitAl.permgroup.intersect_groups — Function

intersect_groups(G, H)

Return generators of G ∩ H using a complete orbit-stabilizer recursion.

Selects a base point α (the largest point moved by either group). For each β in the orbit intersection α^G ∩ α^H, searches the coset G_α · t_β (where t_β ∈ G is the BFS transversal element with α^{t_β} = β) for an element that also lies in H; if found it becomes a coset representative of G ∩ H. The stabilizer part G_α ∩ H_α is found by the same algorithm applied recursively.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3,4]); t = Perm([1,2,4,3]);

julia> G = PermGp([s, t], one(s));  # Klein 4-group

julia> H = PermGp([s], one(s));     # ℤ/2

julia> K = intersect_groups(G, H);

julia> elements(K)
2-element Vector{Perm}:
 Perm([1, 2, 3, 4])
 Perm([2, 1, 3, 4])

intersect_groups(G::SimsGp, H::SimsGp)

Return generators of G ∩ H as a SimsGp, using an orbit-stabilizer recursion with Schreier-Sims membership testing.

Uses orbit_sims to compute the orbit and stabilizer of a base point α in both groups. For each β in α^G ∩ α^H, searches the coset G_α · t_β for an element lying in H, using the Schreier-Sims membership test on H. The stabilizer part G_α ∩ H_α is found by the same algorithm recursively.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3,4]); t = Perm([1,2,4,3]);

julia> G = SimsGp([s, t], one(s));  # Klein 4-group

julia> H = SimsGp([s], one(s));     # ℤ/2

julia> K = intersect_groups(G, H);

julia> size(K)
2