Permutation Groups

This module provides lightweight permutation groups built on top of the orbit algorithms. A group is defined by a list of generators and an identity element; all group-theoretic operations (size, membership, conjugacy classes) are derived from orbit computations.

Types

OrbitAl.permgroup.PermGp — Type

PermGp(gens, one)

A permutation group with generators gens::Vector{Perm} and identity element one.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> G = PermGp([s, t], one(s));

julia> sizeOfGroup(G)
6

Element Enumeration

OrbitAl.permgroup.elements — Function

elements(group)

Return all elements of group as a sorted vector.

OrbitAl.permgroup.subgroups — Function

subgroups(group)

Return all subgroups of group as a vector of PermGp values, computed by orbit enumeration under the closure action.

Size and Membership

OrbitAl.permgroup.sizeOfGroup — Function

sizeOfGroup(group)

Return the order of group using the orbit-stabilizer theorem, without enumerating all elements.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> sizeOfGroup(PermGp([s, t], one(s)))
6

OrbitAl.permgroup.memberOfGroup — Function

memberOfGroup(group, perm)

Return true if perm belongs to group, using a recursive sifting algorithm.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> G = PermGp([s, t], one(s));

julia> memberOfGroup(G, s)
true

julia> memberOfGroup(G, Perm([2,3,4,1]))
false

OrbitAl.permgroup.randomGroupElement — Function

randomGroupElement(group)

Return a uniformly random element of group using a recursive orbit-stabilizer decomposition.

Conjugacy

OrbitAl.permgroup.conjClasses — Function

conjClasses(gp)

Return the conjugacy classes of gp as a vector of Orbit objects, each containing the elements of one class.

OrbitAl.permgroup.subgpClasses — Function

subgpClasses(gp)

Return the conjugacy classes of subgroups of gp as a vector of Orbit objects.

OrbitAl.permgroup.closure — Function

closure(group, a)

Return a new PermGp formed by adding generator a to group.

Subgroup Structure

OrbitAl.permgroup.isPrimePower — Function

isPrimePower(n)

Return true if n > 1 is a power of a single prime, i.e., n = p^k for some prime p and k ≥ 1.

OrbitAl.permgroup.zuppos — Function

zuppos(group)

Return one generator for each cyclic subgroup of prime-power order (zuppo) of group. The returned generators suffice to find all subgroups of group under the closure action.

Intersection

Base.intersect — Method

intersect(G, H)
G ∩ H

Return generators of G ∩ H by a BFS through H's Cayley graph, starting from the identity. A Dict index deduplicates visited elements. For each new element s not yet indexed: if s already lies in the current known intersection K, it is recorded but not expanded; otherwise s is added to the BFS queue and, if s ∈ G, added as a new generator of K. By the coset argument, every element of H not in K when first reached represents a distinct coset of K in H, so at most [H : G∩H] elements enter the queue.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3,4]); t = Perm([1,3,2,4]); u = Perm([1,2,4,3]);

julia> G = PermGp([s, t], one(s));  # S3 on {1,2,3}, order 6

julia> H = PermGp([t, u], one(s));  # S3 on {2,3,4}, order 6

julia> K = G ∩ H;

julia> elements(K)
2-element Vector{Perm}:
 Perm([1, 2, 3, 4])
 Perm([1, 3, 2, 4])