Permutation Groups

This module provides lightweight permutation groups built on top of the orbit algorithms. A group is defined by a list of generators and an identity element; all group-theoretic operations (size, membership, conjugacy classes) are derived from orbit computations.

Types

OrbitAl.permgroup.PermGp — Type

PermGp(gens, one)

A permutation group with generators gens::Vector{Perm} and identity element one.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> G = PermGp([s, t], one(s));

julia> sizeOfGroup(G)
6

Element Enumeration

OrbitAl.permgroup.elements — Function

elements(group)

Return all elements of group as a sorted vector.

OrbitAl.permgroup.subgroups — Function

subgroups(group)

Return all subgroups of group as a vector of PermGp values, computed by orbit enumeration under the closure action.

Size and Membership

OrbitAl.permgroup.sizeOfGroup — Function

sizeOfGroup(group)

Return the order of group using the orbit-stabilizer theorem, without enumerating all elements.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> sizeOfGroup(PermGp([s, t], one(s)))
6

OrbitAl.permgroup.memberOfGroup — Function

memberOfGroup(group, perm)

Return true if perm belongs to group, using a recursive sifting algorithm.

Examples

julia> using OrbitAl

julia> s = Perm([2,1,3]); t = Perm([1,3,2]);

julia> G = PermGp([s, t], one(s));

julia> memberOfGroup(G, s)
true

julia> memberOfGroup(G, Perm([2,3,4,1]))
false

OrbitAl.permgroup.randomGroupElement — Function

randomGroupElement(group)

Return a uniformly random element of group using a recursive orbit-stabilizer decomposition.

Conjugacy

OrbitAl.permgroup.conjClasses — Function

conjClasses(gp)

Return the conjugacy classes of gp as a vector of Orbit objects, each containing the elements of one class.

OrbitAl.permgroup.subgpClasses — Function

subgpClasses(gp)

Return the conjugacy classes of subgroups of gp as a vector of Orbit objects.

OrbitAl.permgroup.closure — Function

closure(group, a)

Return a new PermGp formed by adding generator a to group.